% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bias.R
\name{compute_bias_25A}
\alias{compute_bias_25A}
\title{Compute bias using scoringutils}
\usage{
compute_bias_25A(df_prepared, locs, nowcast_dates)
}
\arguments{
\item{df_prepared}{Prepared data in long format}

\item{locs}{Vector of location codes to include}

\item{nowcast_dates}{Vector of nowcast dates to include}
}
\value{
Data frame with bias scores by model, location, and nowcast_date
}
\description{
Compute bias using scoringutils
}
