% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_coverage.R
\name{compute_coverage}
\alias{compute_coverage}
\title{Compute prediction interval coverage using scoringutils}
\usage{
compute_coverage(df_prepared, locs, nowcast_dates, intervals = c(50, 95))
}
\arguments{
\item{df_prepared}{Prepared data in long format}

\item{locs}{Vector of location codes to include}

\item{nowcast_dates}{Vector of nowcast dates to include}

\item{intervals}{Numeric vector of interval ranges to include
(default: c(50, 95)}
}
\value{
Data frame with coverage for 50\% and 95\% intervals by model,
location, and nowcast_date
}
\description{
Compute prediction interval coverage using scoringutils
}
