% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{convert_to_su_object}
\alias{convert_to_su_object}
\title{Convert the scores to a scoringutils object}
\usage{
convert_to_su_object(
  scores_data,
  brier_to_use = c("brier_point", "brier_dist")
)
}
\arguments{
\item{scores_data}{Data.frame from Variant Nowcast Hub GitHub}

\item{brier_to_use}{Character string indicating}
}
\value{
scoringutils object
}
\description{
Convert the scores to a scoringutils object
}
