% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{create_per_model_heatmap_fig}
\alias{create_per_model_heatmap_fig}
\title{Create per-model submission heatmap figure}
\usage{
create_per_model_heatmap_fig(
  submission_data,
  seq_counts_by_loc,
  plot_components,
  output_fp = file.path("output", "figs", "metadata", "supp"),
  plot_name = "submission_heatmap_by_model"
)
}
\arguments{
\item{submission_data}{Prepared submission status data}

\item{plot_components}{List containing theme and color information}

\item{output_fp}{Character string indicating directory to save file}

\item{plot_name}{Character string indicating name of plot}
}
\value{
Combined patchwork plot of all per-model heatmaps
}
\description{
Create per-model submission heatmap figure
}
