% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{create_summary_heatmap_figure}
\alias{create_summary_heatmap_figure}
\title{Create summary submission heatmap figure}
\usage{
create_summary_heatmap_figure(submission_data, plot_components)
}
\arguments{
\item{submission_data}{Prepared submission status data}

\item{plot_components}{List containing theme and color information}
}
\value{
ggplot2 object showing summary across all models
}
\description{
Create summary submission heatmap figure
}
