% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_weekly.R
\name{daily_to_weekly}
\alias{daily_to_weekly}
\title{Convert sequence data from daily to weekly}
\usage{
daily_to_weekly(daily_data)
}
\arguments{
\item{daily_data}{Data.frame of sequence counts by location and clade}
}
\value{
Data aggregated by epiweek.
}
\description{
Convert sequence data from daily to weekly
}
