% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_weekly.R
\name{daily_to_weekly_scores}
\alias{daily_to_weekly_scores}
\title{Convert scores data from daily to weekly}
\usage{
daily_to_weekly_scores(daily_data)
}
\arguments{
\item{daily_data}{Data.frame of scores by location, model, and nowcast date}
}
\value{
Data aggregated by epiweek.
}
\description{
Convert scores data from daily to weekly
}
