% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{extract_nowcasts}
\alias{extract_nowcasts}
\title{Extract nowcasts from S3 bucket}
\usage{
extract_nowcasts(
  nowcast_dates,
  states,
  bucket_name,
  clades_w_display_name = clades_w_display_name
)
}
\arguments{
\item{nowcast_dates}{Nowcast date(s) to extract}

\item{states}{Locations(s) to extract}

\item{bucket_name}{Name of the S3 bucket}
}
\value{
Raw nowcast outputs as a table for the selected locations and
nowcast dates
}
\description{
Extract nowcasts from S3 bucket
}
