% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_figs.R
\name{get_bar_chart_clade_comp}
\alias{get_bar_chart_clade_comp}
\title{Get a plot of the weekly observed clade frequency in CA dataset compared
to in the NCBI dataset}
\usage{
get_bar_chart_clade_comp(
  clade_comp_weekly,
  fig_fp = file.path("output", "figs", "supp")
)
}
\arguments{
\item{clade_comp_weekly}{Dataframe with counts of number of sequences by
clade and nowcast date for both data sources}

\item{fig_fp}{Character string indicating filepath of figure}
}
\value{
ggplot of weekly clade frequencies for both data sources
}
\description{
Get a plot of the weekly observed clade frequency in CA dataset compared
to in the NCBI dataset
}
