% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data_as_of.R
\name{get_bar_chart_comparison}
\alias{get_bar_chart_comparison}
\title{Bar chart of sequence counts stacked and colored by clade}
\usage{
get_bar_chart_comparison(
  obs_data,
  final_data,
  location,
  temporal_granularity,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs"),
  log_scale = TRUE
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{final_data}{Data.frame of final sequences by clade and location}

\item{location}{Location to plot (abbreviation)}

\item{temporal_granularity}{Temporal granularity to plot}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}

\item{log_scale}{Boolean indicating whether or not y axis should be on log
scale, default is TRUE}

\item{nowcast_date_line}{Boolean indicating whether or not to include a
dashed line for the nowcast date}
}
\value{
ggplot object
}
\description{
Bar chart of sequence counts stacked and colored by clade
}
