% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_bar_chart_seq_count}
\alias{get_bar_chart_seq_count}
\title{Bar chart of sequence counts stacked and colored by clade}
\usage{
get_bar_chart_seq_count(
  obs_data,
  location,
  temporal_granularity,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs")
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{location}{Location to plot (abbreviation)}

\item{temporal_granularity}{Temporal granularity to plot}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
ggplot object
}
\description{
Bar chart of sequence counts stacked and colored by clade
}
