% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_by_loc_figure}
\alias{get_by_loc_figure}
\title{Get a plot of the summary of the overall scores}
\usage{
get_by_loc_figure(
  a,
  b,
  c,
  d,
  e,
  output_fp = file.path("output", "figs", "overall_scores"),
  plot_name = "by_location"
)
}
\arguments{
\item{a}{A}

\item{b}{B}

\item{c}{C}

\item{d}{D}

\item{e}{E}

\item{output_fp}{directory to save figures}

\item{plot_name}{name of the plot}
}
\value{
ggplot object
}
\description{
Get a plot of the summary of the overall scores
}
