% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_d_energy_by_date}
\alias{get_chart_d_energy_by_date}
\title{Chart D: Energy score skill by nowcast date (half-width panel)}
\usage{
get_chart_d_energy_by_date(scores_obj)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}
}
\value{
ggplot object
}
\description{
Chart D: Energy score skill by nowcast date (half-width panel)
}
