% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_e_skill_by_location}
\alias{get_chart_e_skill_by_location}
\title{Chart E: Overall skill by location with model color-coding}
\usage{
get_chart_e_skill_by_location(
  scores_obj,
  score_type = c("brier_score", "energy_score")
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score to plot}
}
\value{
ggplot object
}
\description{
Chart E: Overall skill by location with model color-coding
}
