% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_f_weekly_seq_by_location}
\alias{get_chart_f_weekly_seq_by_location}
\title{Chart F: Weekly sequence averages by location}
\usage{
get_chart_f_weekly_seq_by_location(variant_data)
}
\arguments{
\item{variant_data}{Data frame with variant sequence data}
}
\value{
ggplot object
}
\description{
Chart F: Weekly sequence averages by location
}
