% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_g_timeseries_all_regions}
\alias{get_chart_g_timeseries_all_regions}
\title{Chart G: Time-series skill trends for all US regions}
\usage{
get_chart_g_timeseries_all_regions(
  scores_obj,
  score_type = c("brier_score", "energy_score")
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score to plot}
}
\value{
ggplot object
}
\description{
Chart G: Time-series skill trends for all US regions
}
