% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_h_timeseries_california}
\alias{get_chart_h_timeseries_california}
\title{Chart H: Time-series skill trends for California}
\usage{
get_chart_h_timeseries_california(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  location = "CA"
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score to plot}

\item{location}{Character string indicating location (default "CA")}
}
\value{
ggplot object
}
\description{
Chart H: Time-series skill trends for California
}
