% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_chart_i_avg_scores_over_time}
\alias{get_chart_i_avg_scores_over_time}
\title{Chart I: Average scores over time}
\usage{
get_chart_i_avg_scores_over_time(scores_obj)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}
}
\value{
ggplot object
}
\description{
Chart I: Average scores over time
}
