% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_clade_display_name}
\alias{get_clade_display_name}
\title{Get clade display name}
\usage{
get_clade_display_name(
  url =
    "https://raw.githubusercontent.com/nextstrain/ncov/refs/heads/master/defaults/clade_display_names.yml"
)
}
\arguments{
\item{url}{URL to yaml mapping clade to pangolineage clade name}
}
\value{
Tibble with clade and clade with pangolineage name next to it
}
\description{
Get clade display name
}
