% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_clade_list}
\alias{get_clade_list}
\title{Get vector of clades}
\usage{
get_clade_list(nowcast_dates, clades_by_nowcast_date_dir)
}
\arguments{
\item{nowcast_dates}{Vector of nowcast dates to be evaluated}

\item{clades_by_nowcast_date_dir}{Character string indicating the directory
of the clades}
}
\value{
Character string of all clades used during the nowcast dates
}
\description{
Get vector of clades
}
