% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_clean_variant_data}
\alias{get_clean_variant_data}
\title{Get clean variant data from raw data}
\usage{
get_clean_variant_data(
  raw_variant_data,
  clade_list,
  location_data,
  nowcast_dates,
  nowcast_days = 31
)
}
\arguments{
\item{raw_variant_data}{Data.frame of latest data extracted directly from
next strain}

\item{clade_list}{Vector of character strings of the clade names}

\item{location_data}{Data.frame of location information}

\item{nowcast_dates}{Vector of character strings indicate the date range of
the data.}

\item{nowcast_days}{Number of days we nowcast, default is \code{31}.}
}
\value{
Data.frame of counts of sequences of each clade we nowcasted during
the season
}
\description{
Get clean variant data from raw data
}
