% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data_as_of.R
\name{get_data_fig_all_us_horizontal}
\alias{get_data_fig_all_us_horizontal}
\title{Make data comparing data as of nowcast date vs when evaluating -
Horizontal layout}
\usage{
get_data_fig_all_us_horizontal(
  seq_counts_as_of,
  seq_counts_eval,
  eval_freq,
  plot_name,
  output_fp = file.path("output", "figs", "as_of_data_figs", "final")
)
}
\arguments{
\item{seq_counts_as_of}{A plot showing sequence counts as of nowcast date}

\item{seq_counts_eval}{B plot showing sequence counts for evaluation}

\item{eval_freq}{C plot showing clade frequencies}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
patchwork figure
}
\description{
Make data comparing data as of nowcast date vs when evaluating -
Horizontal layout
}
