% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_nowcasts.R
\name{get_faceted_seq_as_of}
\alias{get_faceted_seq_as_of}
\title{Bar chart of sequence counts stacked and colored by clade}
\usage{
get_faceted_seq_as_of(
  obs_data,
  temporal_granularity = "weeks",
  log_scale = TRUE,
  prev_data_to_show = 70,
  remove_legend = TRUE
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{temporal_granularity}{Temporal granularity to plot}

\item{log_scale}{Boolean indicating whether or not y axis should be on log
scale, default is TRUE.}

\item{prev_data_to_show}{Number of days before nowcast date to show data,
default is 70.}

\item{location}{Location to plot (abbreviation)}
}
\value{
ggplot object
}
\description{
Bar chart of sequence counts stacked and colored by clade
}
