% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_nowcasts.R
\name{get_fig_nowcasts}
\alias{get_fig_nowcasts}
\title{Make data comparing data as of nowcast date vs when evaluating}
\usage{
get_fig_nowcasts(
  faceted_nowcasts,
  row_seq_counts,
  plot_name,
  output_fp = file.path("output", "figs", "nowcasts", "final")
)
}
\arguments{
\item{faceted_nowcasts}{A}

\item{row_seq_counts}{B}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
patchwork figure
}
\description{
Make data comparing data as of nowcast date vs when evaluating
}
