% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_first_data_fig}
\alias{get_first_data_fig}
\title{Patchwork first data figure}
\usage{
get_first_data_fig(
  plot_freq,
  plot_seq,
  plot_hosp,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs", "final")
)
}
\arguments{
\item{plot_freq}{A}

\item{plot_seq}{B}

\item{plot_hosp}{C}

\item{plot_name}{name of figure}

\item{output_fp}{directory to save}
}
\value{
combined figure
}
\description{
Patchwork first data figure
}
