% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_first_data_fig_alt}
\alias{get_first_data_fig_alt}
\title{Patchwork first data figure -- alternate version with US mins CA and CA}
\usage{
get_first_data_fig_alt(
  plot_freq1,
  plot_seq1,
  plot_hosp1,
  plot_freq2,
  plot_seq2,
  plot_hosp2,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs", "final")
)
}
\arguments{
\item{plot_freq1}{A}

\item{plot_seq1}{B}

\item{plot_hosp1}{C}

\item{plot_freq2}{D}

\item{plot_seq2}{E}

\item{plot_hosp2}{F}

\item{plot_name}{name of figure}

\item{output_fp}{directory to save}
}
\value{
combined figure
}
\description{
Patchwork first data figure -- alternate version with US mins CA and CA
}
