% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_first_data_fig_triangular}
\alias{get_first_data_fig_triangular}
\title{First data figure - Triangular layout}
\usage{
get_first_data_fig_triangular(
  plot_freq,
  plot_seq,
  plot_hosp,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs", "final")
)
}
\arguments{
\item{plot_freq}{Plot showing observed clade frequencies}

\item{plot_seq}{Plot showing sequence counts}

\item{plot_hosp}{Plot showing hospital admissions}

\item{plot_name}{Name of figure}

\item{output_fp}{Directory to save}
}
\value{
patchwork figure
}
\description{
First data figure - Triangular layout
}
