% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_heatmap_rel_skill_by_model}
\alias{get_heatmap_rel_skill_by_model}
\title{Heatmap of brier/Energy Relative skill}
\usage{
get_heatmap_rel_skill_by_model(
  scores_obj,
  seq_counts_by_date_loc,
  threshold_n_seq = 5,
  plot_name,
  output_fp = file.path("output", "figs", "supp"),
  score_type = c("brier_score", "energy_score")
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{seq_counts_by_date_loc}{Number of sequences at each location and
nowcast date.}

\item{threshold_n_seq}{Minimum number of sequences for a particular nowcast
date and location for the scaled relative skill to be included in the
average}

\item{plot_name}{Name of plot}

\item{output_fp}{directory to save figures}

\item{score_type}{Character string indicating which score metric to use}
}
\value{
ggplot object
}
\description{
Heatmap of brier/Energy Relative skill
}
