% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_oracle_output}
\alias{get_oracle_output}
\title{Extract oracle output data which will be used for evaluation}
\usage{
get_oracle_output(hub_path, nowcast_dates, states = NULL)
}
\arguments{
\item{hub_path}{Character string of hub path url}

\item{nowcast_dates}{Vector of character strings of dates to extract}

\item{states}{Vector of charatcer strings of locations to extract}
}
\value{
Data.frame of sequence counts by location, date, and clade
as of 90 days after each nowcast date, for evalauting that nowcast date.
}
\description{
Extract oracle output data which will be used for evaluation
}
