% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_overall_scores_figure}
\alias{get_overall_scores_figure}
\title{Combine all 9 charts into a comprehensive figure using patchwork}
\usage{
get_overall_scores_figure(
  chart_a,
  chart_b,
  chart_c,
  chart_d,
  chart_e,
  chart_f,
  chart_g,
  chart_h,
  chart_i,
  plot_name = "overall_scores_summary",
  output_fp = file.path("output", "figs", "overall_scores")
)
}
\arguments{
\item{chart_a}{Chart A plot object}

\item{chart_b}{Chart B plot object}

\item{chart_c}{Chart C plot object}

\item{chart_d}{Chart D plot object}

\item{chart_e}{Chart E plot object}

\item{chart_f}{Chart F plot object}

\item{chart_g}{Chart G plot object}

\item{chart_h}{Chart H plot object}

\item{chart_i}{Chart I plot object}

\item{plot_name}{Character string indicating name of output file}

\item{output_fp}{Character string indicating directory to save file}
}
\value{
Combined patchwork plot object
}
\description{
Combine all 9 charts into a comprehensive figure using patchwork
}
