% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_panel_horizon}
\alias{get_panel_horizon}
\title{Get a plot of absolute scores by horizon}
\usage{
get_panel_horizon(
  a,
  b,
  c,
  d,
  plot_name,
  output_fp = file.path("output", "figs", "overall_scores", "supp")
)
}
\arguments{
\item{a}{A}

\item{b}{B}

\item{c}{C}

\item{d}{D}

\item{plot_name}{name of the plot}

\item{output_fp}{directory to save figures}
}
\value{
ggplot object
}
\description{
Get a plot of absolute scores by horizon
}
