% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_avg_rel_skill_by_loc}
\alias{get_plot_avg_rel_skill_by_loc}
\title{Brier/Energy Relative skill averaged across dates by location}
\usage{
get_plot_avg_rel_skill_by_loc(
  scores_obj,
  seq_counts_by_loc,
  plot_name,
  output_fp = file.path("output", "figs", "supp"),
  score_type = c("brier_score", "energy_score"),
  remove_legend = FALSE
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{seq_counts_by_loc}{Total sequences for each location}

\item{plot_name}{Name of plot}

\item{output_fp}{directory to save figures}

\item{score_type}{Character string indicating which score metric to use}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative skill averaged across dates by location
}
