% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_avg_rel_skill_overall}
\alias{get_plot_avg_rel_skill_overall}
\title{Brier/Energy Relative skill averaged by model}
\usage{
get_plot_avg_rel_skill_overall(
  scores_obj,
  seq_counts_by_date_loc,
  plot_name,
  threshold_n_seq = 5,
  output_fp = file.path("output", "figs", "supp"),
  score_type = c("brier_score", "energy_score"),
  remove_legend = FALSE,
  add_shape = FALSE,
  title = NULL
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{seq_counts_by_date_loc}{Number of sequences for evaluation for each
nowcast date, collection date, and location}

\item{plot_name}{Name of plot}

\item{output_fp}{directory to save figures}

\item{score_type}{Character string indicating which score metric to use}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}

\item{add_shape}{Boolean indicating whether to add the shape legend,
default is FALSE.}

\item{title}{Character string indicating title, default is NULL.}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative skill averaged by model
}
