% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_by_avg_rel_skil_by_t}
\alias{get_plot_by_avg_rel_skil_by_t}
\title{Brier/Energy Relative averaged across locations by nowcast date}
\usage{
get_plot_by_avg_rel_skil_by_t(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  rel_skill_plot = TRUE,
  remove_legend = TRUE,
  title = NULL
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score metric to use}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}

\item{title}{Character string indicating title, default is NULL.}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative averaged across locations by nowcast date
}
