% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_by_location}
\alias{get_plot_by_location}
\title{Brier/Energy Relative/Overall by location}
\usage{
get_plot_by_location(
  scores_obj,
  seq_counts_by_loc,
  score_type = c("brier_score", "energy_score"),
  rel_skill_plot = TRUE,
  remove_legend = TRUE
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{seq_counts_by_loc}{Total sequences for each location}

\item{score_type}{Character string indicating which score metric to use}

\item{rel_skill_plot}{Boolean indicating to return a relative skill plot}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative/Overall by location
}
