% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_plot_coverage_by_date}
\alias{get_plot_coverage_by_date}
\title{Get a plot of prediction interval coverage summarized across nowcast dates}
\usage{
get_plot_coverage_by_date(coverage_data, locs, nowcast_dates, date_range)
}
\arguments{
\item{coverage_data}{Data.frame of coverage scores with interval_range}

\item{locs}{Vector of character strings of locations}

\item{nowcast_dates}{Set of nowcast dates to include}

\item{date_range}{Range of dates to plot (not used but kept for compatibility)}
}
\value{
ggplot
}
\description{
Get a plot of prediction interval coverage summarized across nowcast dates
}
