% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_plot_coverage_overall}
\alias{get_plot_coverage_overall}
\title{Get a plot of prediction interval coverage summarized across nowcast dates}
\usage{
get_plot_coverage_overall(coverage, locs)
}
\arguments{
\item{locs}{Vector of character strings of locations}

\item{coverage_data}{Data.frame of coverage scores with interval_range}
}
\value{
ggplot
}
\description{
Get a plot of prediction interval coverage summarized across nowcast dates
}
