% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data_as_of.R
\name{get_plot_eval_days}
\alias{get_plot_eval_days}
\title{Get a plot of the proportion of days evaluated}
\usage{
get_plot_eval_days(su_scores, nowcast_date, date_range)
}
\arguments{
\item{su_scores}{full scoringutils object}

\item{nowcast_date}{Nowcast date to filter}

\item{date_range}{Dates to plot}
}
\value{
ggplot plot
}
\description{
Get a plot of the proportion of days evaluated
}
