% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_nowcasts.R
\name{get_plot_model_outputs}
\alias{get_plot_model_outputs}
\title{Get a plot of the nowcasts for a specific nowcast date}
\usage{
get_plot_model_outputs(
  nowcasts,
  data_as_of,
  plot_name,
  output_fp = file.path("output", "figs", "nowcasts"),
  save = TRUE,
  remove_legend = TRUE,
  title = NULL,
  temporal_granularity = "weeks",
  prev_data_to_show = 70
)
}
\arguments{
\item{nowcasts}{Data.frame with nowcasts for a single nowcast date for
multiple teams and (a few) locations}

\item{data_as_of}{Data.frame with data for a few locations up until a single
nowcast date}

\item{plot_name}{Character string indicating name of plot}

\item{output_fp}{Character string indicating directory to save file.}

\item{save}{Boolean indicating whether or not to save the plot.}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}

\item{title}{Character string indicating title, default is NULL.}
}
\value{
ggplot object
}
\description{
Get a plot of the nowcasts for a specific nowcast date
}
