% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_pred_plus_data.R
\name{get_plot_model_pred_obs}
\alias{get_plot_model_pred_obs}
\title{Plot the model predicted observations vs the weekly frequencies}
\usage{
get_plot_model_pred_obs(model_pred_obs_df, eval_seq, clades_to_plot = NULL)
}
\arguments{
\item{model_pred_obs_df}{Data.frame of model predicted observed frequency
quantile values}

\item{eval_seq}{Counts of sequences by clade for evaluating on a particular
nowcast date}

\item{clades_to_plot}{Vector of chharacter strings of clades to plot,
default is NULL which plots all of them}
}
\value{
ggplot object
}
\description{
Plot the model predicted observations vs the weekly frequencies
}
