% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_plot_model_preds_mult_nowcasts}
\alias{get_plot_model_preds_mult_nowcasts}
\title{Get a plot of nowcasts for multiple nowcast dates vs observations}
\usage{
get_plot_model_preds_mult_nowcasts(
  model_preds_mult_nowcasts,
  final_eval_data,
  clade_to_zoom = "25A",
  horizon_to_plot = c(-6, 0)
)
}
\arguments{
\item{model_preds_mult_nowcasts}{Model predicted observations for multiple
nowcasts}

\item{final_eval_data}{Final sequence counts by clade and location}

\item{clade_to_zoom}{Character string indicating the clade to zoom}

\item{horizon_to_plot}{Range of integers to filter to for the plot}
}
\value{
ggplot object
}
\description{
Get a plot of nowcasts for multiple nowcast dates vs observations
}
