% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_mult_locs}
\alias{get_plot_mult_locs}
\title{Plot multiple locations}
\usage{
get_plot_mult_locs(
  data,
  plot_name,
  output_fp = file.path("output", "figs", "eda"),
  save = TRUE
)
}
\arguments{
\item{data}{Data.frame with all sequence data for multiple locations}
}
\value{
ggplot object
}
\description{
Plot multiple locations
}
