% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_overall}
\alias{get_plot_overall}
\title{Brier/Energy Relative/Overall by model}
\usage{
get_plot_overall(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  rel_skill_plot = TRUE,
  remove_legend = TRUE,
  title = NULL
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score metric to use}

\item{rel_skill_plot}{Boolean indicating to return a relative skill plot}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}

\item{title}{Character string indicating title, default is NULL.}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative/Overall by model
}
