% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_figs.R
\name{get_plot_prop_excl_by_horizon_loc}
\alias{get_plot_prop_excl_by_horizon_loc}
\title{Get a plot of the proportion of days with which partial observations
existed, by horizon, for each location}
\usage{
get_plot_prop_excl_by_horizon_loc(
  su_scores,
  fig_fp = file.path("output", "figs", "supp")
)
}
\arguments{
\item{su_scores}{Scoringutils score object with additional column of metadata
indicating whether or not to score the prediction.}

\item{fig_fp}{Character string indicating filepath of figure}
}
\value{
ggplot of prop of predictions scored by horizons for each location
}
\description{
Get a plot of the proportion of days with which partial observations
existed, by horizon, for each location
}
