% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_rel_skill_overall}
\alias{get_plot_rel_skill_overall}
\title{Get a plot of relative scaled skill overall}
\usage{
get_plot_rel_skill_overall(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  plot_name = "overall_rel_scaled_skill",
  output_fp = file.path("output", "figs", "eda"),
  save = TRUE
)
}
\arguments{
\item{scores_obj}{Scoringutils scores objected}

\item{score_type}{Character string indicating which score to plot.}

\item{plot_name}{Character string indicating name of plot}

\item{output_fp}{Character string indicating directory to save file.}

\item{save}{Boolean indicating whether or not to save the plot.}
}
\value{
ggplot object bar chart of scores
}
\description{
Get a plot of relative scaled skill overall
}
