% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_plot_scores_by_nowcast_date}
\alias{get_plot_scores_by_nowcast_date}
\title{Get a plot of scores by nowcast date for locations and models}
\usage{
get_plot_scores_by_nowcast_date(
  scores,
  locs,
  nowcast_dates,
  date_range,
  horizon_to_plot = c(-6, 0)
)
}
\arguments{
\item{scores}{Data.frame of scores}

\item{locs}{Vector of character strings of locations}

\item{nowcast_dates}{Set of nowcast dates to summarise over}

\item{date_range}{}

\item{horizon_to_plot}{}
}
\description{
Get a plot of scores by nowcast date for locations and models
}
