% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_pred_plus_data.R
\name{get_plot_scores_over_time_bar}
\alias{get_plot_scores_over_time_bar}
\title{Plot scores over time colored by model as a dodged bar chart}
\usage{
get_plot_scores_over_time_bar(scores, locs, nowcast_date)
}
\arguments{
\item{scores}{Data.frame of all scores for all nowcast dates and locations}

\item{locs}{Vector of character strings of locations}

\item{nowcast_date}{Nowcast date to focus on}
}
\value{
ggplot
}
\description{
Plot scores over time colored by model as a dodged bar chart
}
