% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_scores_t}
\alias{get_plot_scores_t}
\title{Plot scores over time for a few locations and a specific nowcast date}
\usage{
get_plot_scores_t(
  scores_data_hub,
  locations,
  this_nowcast_date,
  score_type = c("energy", "brier_point", "brier_dist"),
  plot_name = "scores_over_time",
  output_fp = file.path("output", "figs", "eda"),
  save = TRUE
)
}
\arguments{
\item{scores_data_hub}{Data.frame of scores from the Hub}

\item{locations}{Vector of character strings indicating the set of locations
to plot as facets}

\item{this_nowcast_date}{Character string indicating nowcast date to plot}

\item{score_type}{Character string indicating which score to plot.}

\item{plot_name}{Character string indicating name of plot}

\item{output_fp}{Character string indicating directory to save file.}

\item{save}{Boolean indicating whether or not to save the plot.}
}
\value{
ggplot object
}
\description{
Plot scores over time for a few locations and a specific nowcast date
}
