% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_pred_plus_data.R
\name{get_plot_scores_w_exclusions}
\alias{get_plot_scores_w_exclusions}
\title{Plot scores over time colored by whether or not it was scored}
\usage{
get_plot_scores_w_exclusions(
  scores,
  locs,
  nowcast_date,
  plot_name = "scores_showing_exclusions",
  output_fp = file.path("output", "figs", "pred_obs", "supp")
)
}
\arguments{
\item{scores}{Data.frame of all scores for all nowcast dates and locations}

\item{locs}{Vector of character strings of locations}

\item{nowcast_date}{Nowcast date to focus on}
}
\value{
ggplot
}
\description{
Plot scores over time colored by whether or not it was scored
}
