% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_plot_seq_counts_by_loc}
\alias{get_plot_seq_counts_by_loc}
\title{Get a plot of sequence counts by location for the duration of the nowcast
period}
\usage{
get_plot_seq_counts_by_loc(
  obs_data,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs", "supp")
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
Bar chart of sequence counts by location
}
\description{
Get a plot of sequence counts by location for the duration of the nowcast
period
}
