% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_relative_skill}
\alias{get_relative_skill}
\title{Get relative skill scores with specified grouping}
\usage{
get_relative_skill(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  by = "target_date"
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score metric to use}

\item{by}{Character vector of grouping variables for relative skill calculation}
}
\value{
Data frame with relative skill scores
}
\description{
Get relative skill scores with specified grouping
}
