% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_scores_by_nowcast_date}
\alias{get_scores_by_nowcast_date}
\title{Get a plot of the overall scores by nowcast date}
\usage{
get_scores_by_nowcast_date(
  a,
  b,
  c,
  d,
  e,
  f,
  g,
  h,
  i,
  j,
  output_fp = file.path("output", "figs", "overall_scores"),
  plot_name = "scores_by_nowcast_date"
)
}
\arguments{
\item{a}{A}

\item{b}{B}

\item{c}{C}

\item{d}{D}

\item{e}{E}

\item{f}{F}

\item{g}{G}

\item{h}{H}

\item{i}{I}

\item{j}{J}

\item{output_fp}{directory to save figures}

\item{plot_name}{name of the plot}
}
\value{
ggplot object
}
\description{
Get a plot of the overall scores by nowcast date
}
