% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data_as_of.R
\name{get_second_data_fig}
\alias{get_second_data_fig}
\title{Make data comparing data as of nowcast date vs when evaluating}
\usage{
get_second_data_fig(
  seq_counts_as_of1,
  seq_counts_as_of2,
  seq_counts_eval1,
  seq_counts_eval2,
  eval_freq1,
  eval_freq2,
  prop_eval_days_1,
  prop_eval_days_2,
  plot_name,
  output_fp = file.path("output", "figs", "as_of_data_figs", "final")
)
}
\arguments{
\item{seq_counts_as_of1}{A}

\item{seq_counts_as_of2}{B}

\item{seq_counts_eval1}{C}

\item{seq_counts_eval2}{D}

\item{eval_freq1}{E}

\item{eval_freq2}{F}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
patchwork figure
}
\description{
Make data comparing data as of nowcast date vs when evaluating
}
