% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_helpers.R
\name{get_target_data}
\alias{get_target_data}
\title{Extract target data output so we have what would have been available
to teams as of the nowcast date}
\usage{
get_target_data(hub_path, nowcast_dates, states = NULL)
}
\arguments{
\item{hub_path}{Character string of hub path url}

\item{nowcast_dates}{Vector of character strings of dates to extract}

\item{states}{Vector of character strings of locations to extract}
}
\value{
Data.frame of sequence counts by location, date, and clade
as of the nowcast date.
}
\description{
Extract target data output so we have what would have been available
to teams as of the nowcast date
}
