% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_figs.R
\name{get_volume_prop_comp_weekly}
\alias{get_volume_prop_comp_weekly}
\title{Get a plot of the ratio of number of sequences between the two datasets from
CA and NCBI}
\usage{
get_volume_prop_comp_weekly(
  volume_comp_weekly,
  fig_fp = file.path("output", "figs", "supp")
)
}
\arguments{
\item{volume_comp_weekly}{Dataframe with number of sequences by week and
nowcast date for each dataset}

\item{fig_fp}{Character string indicating filepath of figure}
}
\value{
ggplot of ratio of NCBI to CA data.
}
\description{
Get a plot of the ratio of number of sequences between the two datasets from
CA and NCBI
}
