% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_zip}
\alias{open_zip}
\title{Extract contexts from a zip file}
\usage{
open_zip(url)
}
\arguments{
\item{url}{Link to zip file.}
}
\value{
data in file
}
\description{
Extract contexts from a zip file
}
