% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{plot_eval_sequence_heatmap}
\alias{plot_eval_sequence_heatmap}
\title{Plot evaluation sequence heatmap}
\usage{
plot_eval_sequence_heatmap(
  sequence_data,
  plot_components,
  output_fp = file.path("output", "figs", "metadata", "supp"),
  plot_name = "eval_sequence_counts_heatmap"
)
}
\arguments{
\item{sequence_data}{Prepared sequence count data}

\item{plot_components}{List containing theme and color information}

\item{output_fp}{Character string indicating directory to save file}

\item{plot_name}{Character string indicating name of plot}
}
\value{
ggplot2 object
}
\description{
Plot evaluation sequence heatmap
}
