% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{plot_model_submission_heatmap}
\alias{plot_model_submission_heatmap}
\title{Plot submission heatmap for a single model}
\usage{
plot_model_submission_heatmap(submission_data, model_id, plot_components)
}
\arguments{
\item{submission_data}{Prepared submission status data}

\item{model_id}{Model ID to plot}

\item{plot_components}{List containing theme and color information}
}
\value{
ggplot2 object
}
\description{
Plot submission heatmap for a single model
}
