% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{plot_obs_clade_freq}
\alias{plot_obs_clade_freq}
\title{Plot observed clade frequency}
\usage{
plot_obs_clade_freq(
  obs_data,
  location,
  temporal_granularity,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs")
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{location}{Location to plot}

\item{temporal_granularity}{Temporal granularity to plot}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}
}
\value{
ggplot object
}
\description{
Plot observed clade frequency
}
