% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{plot_submission_summary}
\alias{plot_submission_summary}
\title{Plot summary heatmap across all models}
\usage{
plot_submission_summary(submission_data, plot_components)
}
\arguments{
\item{submission_data}{Prepared submission status data}

\item{plot_components}{List containing theme and color information}
}
\value{
ggplot2 object
}
\description{
Plot summary heatmap across all models
}
