% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_submission_heatmaps.R
\name{prepare_eval_sequence_data}
\alias{prepare_eval_sequence_data}
\title{Prepare evaluation sequence count data}
\usage{
prepare_eval_sequence_data(clean_variant_data_for_eval)
}
\arguments{
\item{clean_variant_data_for_eval}{Cleaned evaluation variant data}
}
\value{
Data frame with total sequences by nowcast_date and location
}
\description{
Prepare evaluation sequence count data
}
