% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_bar_chart_seq_count}
\alias{get_bar_chart_seq_count}
\title{Bar chart of sequence counts stacked and colored by clade}
\usage{
get_bar_chart_seq_count(
  obs_data,
  location,
  temporal_granularity,
  plot_name,
  date_range,
  output_fp = file.path("output", "figs", "data_figs"),
  log_scale = FALSE,
  nowcast_date_line = FALSE,
  title = NULL,
  remove_xticks = FALSE
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{location}{Location to plot (abbreviation)}

\item{temporal_granularity}{Temporal granularity to plot}

\item{plot_name}{Name of plot}

\item{output_fp}{File.path to save plot}

\item{log_scale}{Boolean indicating whether or not y axis should be on log
scale, default is TRUE}

\item{nowcast_date_line}{Boolean indicating whether or not to include a
dashed line for the nowcast date}

\item{title}{Title to plot, default is NULL which plots no title.}

\item{remove_x_ticks}{Boolean indicating to remove x tick text, default is
FALSE}
}
\value{
ggplot object
}
\description{
Bar chart of sequence counts stacked and colored by clade
}
