% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{get_clean_variant_data}
\alias{get_clean_variant_data}
\title{Get clean variant data from raw data}
\usage{
get_clean_variant_data(
  raw_variant_data,
  clade_list,
  location_data,
  nowcast_date,
  seq_col_name,
  type
)
}
\arguments{
\item{raw_variant_data}{Data.frame of latest data extracted directly from
next strain}

\item{clade_list}{Vector of character strings of the clade names}

\item{location_data}{Data.frame of location information}

\item{nowcast_date}{Character string indicate the date of the nowcast.}

\item{seq_col_name}{Character string indicating the name of the column for
number of sequences of that clade}

\item{type}{Character string indicating data is as of the nowcast date or
evaluation data}
}
\value{
Data.frame of counts of sequences of each clade we nowcasted during
the season
}
\description{
Get clean variant data from raw data
}
