% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_fig_zoom_25A}
\alias{get_fig_zoom_25A}
\title{Multiplot panel looking at 25A emergence across nowcast dates}
\usage{
get_fig_zoom_25A(
  grid,
  scores,
  coverage,
  plot_name,
  output_fp = file.path("output", "figs", "zoom_25A", "final")
)
}
\arguments{
\item{grid}{Model predictions plot}

\item{scores}{Energy scores plot}

\item{coverage}{Prediction interval coverage plot}

\item{plot_name}{name of plot}

\item{output_fp}{filepath directory}
}
\value{
patchwork
}
\description{
Multiplot panel looking at 25A emergence across nowcast dates
}
