% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_overall_scores_figure}
\alias{get_overall_scores_figure}
\title{Get a plot of the summary of the overall scores}
\usage{
get_overall_scores_figure(
  a,
  b,
  c,
  d,
  e,
  f,
  g,
  h,
  i,
  j,
  k,
  l,
  output_fp = file.path("output", "figs", "overall_scores"),
  plot_name = "overall_scores"
)
}
\arguments{
\item{a}{A}

\item{b}{B}

\item{c}{C}

\item{d}{D}

\item{e}{E}

\item{f}{F}

\item{g}{G}

\item{h}{H}

\item{i}{I}

\item{j}{J}

\item{k}{K}

\item{l}{L}

\item{output_fp}{directory to save figures}

\item{plot_name}{name of the plot}
}
\value{
ggplot object
}
\description{
Get a plot of the summary of the overall scores
}
