% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_zoom_clade_mult_nowcasts.R
\name{get_plot_bias_by_date}
\alias{get_plot_bias_by_date}
\title{Get a plot of bias by nowcast date for locations and models}
\usage{
get_plot_bias_by_date(
  bias_data,
  locs,
  nowcast_dates,
  date_range,
  plot_name = "bias_over_time_25A",
  output_fp = file.path("output", "figs", "zoom_25A", "supp")
)
}
\arguments{
\item{bias_data}{Data.frame of bias scores}

\item{locs}{Vector of character strings of locations}

\item{nowcast_dates}{Set of nowcast dates to include}

\item{date_range}{Range of dates to plot}

\item{plot_name}{name of plot}

\item{output_fp}{filepath directory}
}
\value{
ggplot
}
\description{
Get a plot of bias by nowcast date for locations and models
}
