% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data_as_of.R
\name{get_plot_freq_as_of_vs_eval}
\alias{get_plot_freq_as_of_vs_eval}
\title{Get a plot of the variant frequency as of the nowcast date versus
once we are ready to evaluate}
\usage{
get_plot_freq_as_of_vs_eval(
  obs_data,
  final_data,
  location,
  temporal_granularity
)
}
\arguments{
\item{obs_data}{Data.frame of number of sequences by clade and location}

\item{final_data}{Data.frame of final sequences by clade and location}

\item{location}{Location to plot (abbreviation)}

\item{temporal_granularity}{Temporal granularity to plot}
}
\value{
plot of clade frequencies as of nowcats date and when we evaluate
}
\description{
Get a plot of the variant frequency as of the nowcast date versus
once we are ready to evaluate
}
