% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_horizon}
\alias{get_plot_horizon}
\title{Brier/Energy Relative/Absolute by horizon}
\usage{
get_plot_horizon(
  scores_obj,
  score_type = c("brier_score", "energy_score"),
  rel_skill_plot = TRUE
)
}
\arguments{
\item{scores_obj}{Scoringutils scores object}

\item{score_type}{Character string indicating which score metric to use}

\item{rel_skill_plot}{Boolean indicating to return a relative skill plot}
}
\value{
ggplot object
}
\description{
Brier/Energy Relative/Absolute by horizon
}
