% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_data.R
\name{get_plot_hosp_admissions}
\alias{get_plot_hosp_admissions}
\title{Percent of hospital admissions due to covid over time}
\usage{
get_plot_hosp_admissions(
  location_to_plot,
  date_range,
  temporal_granularity = "weeks",
  location_data,
  plot_name,
  output_fp = file.path("output", "figs", "data_figs"),
  data_fp = file.path("input", "nssp_states_and_national.csv")
)
}
\arguments{
\item{date_range}{Vector of date range to plot, will use min and max}

\item{temporal_granularity}{Temporal granularity to plot}

\item{location_data}{Data.frame of location metadata to translate codes to
abbreviations}

\item{plot_name}{Name of plot}

\item{output_fp}{File path to save plot}

\item{data_fp}{Filepath to NSSP data at the state and national level}

\item{location}{Location to plot (abbreviation)}
}
\value{
ggplot object
}
\description{
Percent of hospital admissions due to covid over time
}
