% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_nowcasts.R
\name{get_plot_model_outputs}
\alias{get_plot_model_outputs}
\title{Get a plot of the nowcasts for a specific nowcast date}
\usage{
get_plot_model_outputs(
  nowcasts,
  data_as_of,
  remove_legend = TRUE,
  title = NULL,
  temporal_granularity = "weeks",
  prev_data_to_show = 70
)
}
\arguments{
\item{nowcasts}{Data.frame with nowcasts for a single nowcast date for
multiple teams and (a few) locations}

\item{data_as_of}{Data.frame with data for a few locations up until a single
nowcast date}

\item{remove_legend}{Boolean indicating whether to keep legend, default
is TRUE.}

\item{title}{Character string indicating title, default is NULL.}

\item{temporal_granularity}{Temporal granularity to plot}

\item{prev_data_to_show}{Number of days before nowcast date to show data,
default is 70.}
}
\value{
ggplot object
}
\description{
Get a plot of the nowcasts for a specific nowcast date
}
