% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_plots.R
\name{get_plot_nowcasts}
\alias{get_plot_nowcasts}
\title{Get a plot of the nowcasts for a specific nowcast date}
\usage{
get_plot_nowcasts(
  data,
  plot_name,
  output_fp = file.path("output", "figs", "eda"),
  save = TRUE
)
}
\arguments{
\item{data}{Data.frame with nowcasts for a single nowcast date for
multiple teams and (a few) locations}

\item{plot_name}{Character string indicating name of plot}

\item{output_fp}{Character string indicating directory to save file.}

\item{save}{Boolean indicating whether or not to save the plot.}
}
\value{
ggplot object
}
\description{
Get a plot of the nowcasts for a specific nowcast date
}
