% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fig_overall_scores.R
\name{get_plot_seq_counts_date}
\alias{get_plot_seq_counts_date}
\title{Sequence counts by location}
\usage{
get_plot_seq_counts_date(seq_counts_by_date)
}
\arguments{
\item{seq_counts_by_loc}{Total sequences for each location}
}
\value{
ggplot object
}
\description{
Sequence counts by location
}
